This was part of the main part

+++ Project0/BusinessAccount.cs
@@ -6,7 +6,7 @@ namespace Project0
 {
     class BusinessAccount : IAccount
     {
-        public decimal InterestRate { get; set; }
+        public double InterestRate { get; set; }
         public decimal Funds { get; set; }
         public int AccountNumber { get; set; }
         public void Deposit(decimal amt)
